Fixes ghost newlines and moves thermal units to another file
