+++ src/main/kotlin/org/ageseries/libage/sim/Units.kt
@@ -26,6 +26,7 @@ class Scale(
 ) {
     /** Given a value [u] in base unit, return its value in this unit. */
     fun map(u: Double): Double = factor * u + base
+
     /** Given a value [u] in this unit, return its value in base units. */
     fun unmap(u: Double): Double = (u - base) / factor
 
@@ -33,46 +34,4 @@ class Scale(
      * Give a human-readable string of a value in this scale.
      */
     fun display(u: Double): String = "${u}${displayUnit}"
-
-    /**
-     * Linear scales for temperatures.
-     *
-     * All scales are defined relative to Kelvin, which is the package-wide base unit.
-     */
-    object Thermal {
-        /**
-         * Kelvin; absolute zero, with steps of a hundredth of a grade.
-         *
-         * Useless directly, but maybe useful as an argument where a Scale is expected.
-         */
-        val KELVIN = Scale(1.0, 0.0, "K")
-        /**
-         * Absolute Grade; absolute zero, with steps equal to the grade.
-         */
-        val ABSOLUTE_GRADE = Scale(0.01, 0.0, "AG")
-        /**
-         * Rankine; absolute zero, with steps one-hundred-eightieth of a grade.
-         */
-        val RANKINE = Scale(9.0 / 5.0, 0.0, "°R")
-        /**
-         * Grade; zero is pure water freezing, one is pure water boiling, all at standard pressure.
-         */
-        val GRADE = Scale(0.01, -2.7315, "G")
-        /**
-         * Centigrade; zero as Grade, but with steps one-hundredth of a grade.
-         */
-        val CENTIGRADE = Scale(1.0, -273.15, "°C")
-        /**
-         * Celsius, an alias for Centigrade.
-         */
-        val CELSIUS = CENTIGRADE
-        /**
-         * Milligrade; zero as Grade, but with steps one-thousandth of a grade.
-         */
-        val MILLIGRADE = Scale(10.0, -2731.5, "mG")
-        /**
-         * Fahrenheit; 32 is zero grade, 212 is one grade.
-         */
-        val FAHRENHEIT = Scale(9.0 / 5.0, -491.67, "°F")
-    }
 }
